Create Decoder with layers
This is our decoder. After extracting the first 23 layers of the pre-trained mode. we want to add our own decoded layers that are tailored for the purpose of image matting.

Decoder contains 5 Convolutional layers, 4 ReLU layers, 3 Upsample layers, and finishes off with a single sigmoid layer.

The sigmoid layer is best compatible with our loss function.